requirements.txt
py-clob-client
requests
websockets
numpy
pandas
python-dotenv
